requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.25.0
atlastk==0.13
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
et-xmlfile==1.1.0
fastapi==0.78.0
frozenlist==1.3.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pip-autoremove==0.10.0
pycodestyle==2.8.0
pydantic==1.9.1
pypika-tortoise==0.1.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tortoise-orm==0.19.1
typing_extensions==4.2.0
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
yarl==1.7.2
